set backbone float32

+++ src/binaryalign/models/backbone.py
@@ -1,8 +1,10 @@
+import torch
 from transformers import AutoModel
 
 
 def load_backbone(model_name: str, vocab_size: int) -> AutoModel:
-    model = AutoModel.from_pretrained(model_name, use_safetensors=True)
+    
+    model = AutoModel.from_pretrained(model_name, use_safetensors=True, dtype=torch.float32)
 
     model_vocab_size = model.get_input_embeddings().weight.shape[0]
     if model_vocab_size != vocab_size:
